Fix decorator options mutating (fixes #55)
When a class has n properties using the same subclass, the
function `baseProp` is called n times for each property in subclass.

The function `baseProp` receives a parameter `rawOptions` which
should be copied as its being mutated later on. This is intended with
the statement:
```
rawOptions = Object.assign(rawOptions, {})
```
Although, this statement never changes the reference of `rawOptions`
with as result that the options are mutating. This causes the final schema
to be constructed without the options ref, itemsRef, refPath, etc when
the subclass is used more than once.

This commit fixes the mutation of the options.

Signed-off-by: Chris Lahaye <dev@chrislahaye.com>