feat: add support for declaring search indexes
MongoDB and `mongoose` now support adding search indexes to collections.
Introduce a new decorator `@searchIndex()` that, similar to `@index`,
adds a search index to a collection.

Search index support was added to `mongoose` in version `8.1.0`, however
the type definition for `Schema.searchIndex()` didn't exist until
`8.2.3`. Therefore in order to support this feature, update the
`mongoose` dependencies to `~8.2.3`.